+++ src/main/java/com/ratiose/testtask/service/ActorService.java
@@ -4,6 +4,6 @@ import com.ratiose.testtask.entity.Actor;
 import com.ratiose.testtask.entity.User;
 
 public interface ActorService {
-	Actor addFavoritActor(Long favoritActorId, User user);
-	Actor removeFavoriteActor(Long favoriteActorId, User user);	
+	Actor addFavoritActor(Integer favoritActorId, User user);
+	Actor removeFavoriteActor(Integer favoriteActorId, User user);	
 }
